imwrite no cook to float32 & finish examples/image_io.py
