+++ python/taichi/misc/gui.py
@@ -49,6 +49,12 @@ class GUI:
     import taichi as ti
     from .image import cook_image
     img = cook_image(img)
+    if img.dtype in [np.uint8, np.uint16, np.uint32, np.uint64]:
+      img = img.astype(np.float32) * (1 / np.iinfo(img.dtype).max)
+    elif img.dtype in [np.float32, np.float64]:
+      img = img.astype(np.float32)
+    else:
+      raise ValueError(f'Data type {img.dtype} not supported in GUI.set_image')
     if len(img.shape) == 2:
       img = img[..., None]
     if img.shape[2] == 1:
